refactor: Move controllers into _cluster directory
