+++ pkg/clients/_cluster/mock/applications/mock.go
