+++ pkg/clients/_cluster/mock/generate.go
